+++ internal/cachecontroller/cache_controller.go
@@ -142,7 +142,7 @@ func (c *InMemoryCacheController) DetermineInvalidation(
 	return time.Time{}
 }
 
-func (c *InMemoryCacheController) findChanges(ctx context.Context, storeID string) ([]*openfgav1.TupleChange, string, error) {
+func (c *InMemoryCacheController) findChangesDescending(ctx context.Context, storeID string) ([]*openfgav1.TupleChange, string, error) {
 	opts := storage.ReadChangesOptions{
 		SortDesc: true,
 		Pagination: storage.PaginationOptions{
@@ -164,7 +164,8 @@ func (c *InMemoryCacheController) findChangesAndInvalidate(ctx context.Context, 
 	// TODO: this should have a deadline since it will hold up everything if it doesn't return
 	// could also be implemented as a fire and forget mechanism and subsequent requests can grab the result
 	// re-evaluate at a later time.
-	changes, _, err := c.findChanges(ctx, storeID)
+	// Note that changes are sorted most-recent first
+	changes, _, err := c.findChangesDescending(ctx, storeID)
 	if err != nil {
 		telemetry.TraceError(span, err)
 		// do not allow any cache read until next refresh
@@ -172,19 +173,31 @@ func (c *InMemoryCacheController) findChangesAndInvalidate(ctx context.Context, 
 		return
 	}
 
+	mostRecentChanges := changes[0]
 	entry := &storage.ChangelogCacheEntry{
-		LastModified: changes[0].GetTimestamp().AsTime(),
+		LastModified: mostRecentChanges.GetTimestamp().AsTime(),
 	}
 
-	lastInvalidationOccurred := c.cache.Get(cacheKey) != nil
+	lastCacheRecord := c.cache.Get(cacheKey)
 
 	// set changelog entry as soon as possible for subsequent cache
 	// lookups have the entry and not have to wait on the existing singleflight group
 	c.cache.Set(cacheKey, entry, c.ttl)
 
-	timestampOfLastInvalidation := time.Now().Add(-c.ttl)
+	timestampOfLastInvalidation := time.Time{}
+	if lastCacheRecord != nil {
+		decodedRecord, ok := lastCacheRecord.(*storage.ChangelogCacheEntry)
+		if ok {
+			// if the change log cache is available and valid, use the last modified
+			// time to have better consistency. Otherwise, the timestampOfLastInvalidation will
+			// be the beginning of time which imply the need to invalidate one or more records.
+			timestampOfLastInvalidation = decodedRecord.LastModified
+		} else {
+			c.logger.Error("Unable to cast lastCacheRecord properly", zap.String("cacheKey", cacheKey))
+		}
+	}
 
-	if lastInvalidationOccurred && entry.LastModified.Before(timestampOfLastInvalidation) {
+	if entry.LastModified.Before(timestampOfLastInvalidation) {
 		// no new changes, no need to perform invalidations
 		span.SetAttributes(attribute.Bool("invalidations", false))
 		c.logger.Debug("InMemoryCacheController findChangesAndInvalidate invalidation as entry.LastModified before last verified",
@@ -196,11 +209,20 @@ func (c *InMemoryCacheController) findChangesAndInvalidate(ctx context.Context, 
 		return
 	}
 
+	timestampOfLastIteratorInvalidation := time.Now().Add(-c.iteratorCacheTTL)
+
 	// need to consider there might just be 1 change
 	// iterate from the oldest to most recent to determine if the last change is part of the current batch
+	// Remember that idx[0] is the most recent change while idx[len(changes)-1] is the oldest change because
+	// changes is ordered from most recent to oldest.
 	idx := len(changes) - 1
 	for ; idx >= 0; idx-- {
-		if !lastInvalidationOccurred || changes[idx].GetTimestamp().AsTime().After(timestampOfLastInvalidation) {
+		// idx marks the changes the first change after the timestampOfLastIteratorInvalidation.
+		// therefore, we want to use the changes happens at/after this time to invalidate cache.
+		//
+		// Note that we only want to add invalidation entries for changes with timestamp >= now - iterator cache's TTL
+		// because anything older than that time would not live in the iterator cache anyway.
+		if changes[idx].GetTimestamp().AsTime().After(timestampOfLastIteratorInvalidation) {
 			break
 		}
 	}
@@ -212,7 +234,7 @@ func (c *InMemoryCacheController) findChangesAndInvalidate(ctx context.Context, 
 		partialInvalidation = false
 		c.invalidateIteratorCache(storeID)
 	} else {
-		// only a subset of changes are new, revoke the respective ones
+		// only a subset of changes are new, revoke the respective ones.
 		lastModified := time.Now()
 		for ; idx >= 0; idx-- {
 			t := changes[idx].GetTupleKey()
@@ -224,7 +246,7 @@ func (c *InMemoryCacheController) findChangesAndInvalidate(ctx context.Context, 
 	c.logger.Debug("InMemoryCacheController findChangesAndInvalidate invalidation",
 		zap.String("store_id", storeID),
 		zap.Time("entry.LastModified", entry.LastModified),
-		zap.Time("timestampOfLastInvalidation", timestampOfLastInvalidation),
+		zap.Time("timestampOfLastIteratorInvalidation", timestampOfLastIteratorInvalidation),
 		zap.Bool("partialInvalidation", partialInvalidation))
 	span.SetAttributes(attribute.Bool("invalidations", true))
 	findChangesAndInvalidateHistogram.WithLabelValues("true", utils.Bucketize(uint(len(changes)), c.changelogBuckets)).Observe(float64(time.Since(start).Milliseconds()))